+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - complete refactor of Heroku app creation and configuration
 - ask for user credentials before creating the Heroku app
 - set the correct timezone config on heroku after creating the app
+- move RVM config from .rvmrc to .ruby-version and .ruby-gemset files
 
 ### bug fixes